Move automatic handling of LiveStatusTagColumn to IndexViewOptionalFeaturesMixin
This allows us to avoid inspecting the index view class's list_display
in the viewsets, which in turn allows us to use cached property on the
view.
